Refactor totplist.go: split into list/item files and extract shared helper
Split the 411-line totplist.go into two focused files:
- totplist.go (~175 lines): list container management, refresh logic,
  QR scan workflow, and shared showEntryFormDialog helper
- totplist_item.go (~210 lines): individual item rendering and actions
  (create/update item, copy code, entry menu, edit/QR/delete dialogs)

Extract showEntryFormDialog as a common helper to eliminate duplicated
form dialog logic between showEditDialog and showAddConfirmDialog.

Move generateQRCodeImage from ui/qrcode.go to qrscanner package as
an exported GenerateQRCodeImage function.

Closes #7
