+++ src/usecase/qrscanner/generater.go
@@ -1,4 +1,4 @@
-package ui
+package qrscanner
 
 import (
 	"image"
@@ -6,8 +6,8 @@ import (
 	qrcode "github.com/skip2/go-qrcode"
 )
 
-// generateQRCodeImage はURIからQRコード画像を生成する
-func generateQRCodeImage(uri string) (image.Image, error) {
+// GenerateQRCodeImage はURIからQRコード画像を生成する
+func GenerateQRCodeImage(uri string) (image.Image, error) {
 	qr, err := qrcode.New(uri, qrcode.Medium)
 	if err != nil {
 		return nil, err
